adain_cycle_reward/requirements.txt
accelerate==1.12.0
certifi==2025.11.12
charset-normalizer==3.4.4
dreamsim==0.2.1
filelock==3.20.0
fsspec==2025.10.0
ftfy==6.3.1
hf-xet==1.2.0
huggingface-hub==0.36.0
idna==3.11
ImageIO==2.37.2
Jinja2==3.1.6
lazy_loader==0.4
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
open_clip_torch==3.2.0
opencv-python-headless==4.12.0.88
packaging==25.0
peft==0.18.0
pillow==12.0.0
psutil==7.1.3
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
safetensors==0.7.0
scikit-image==0.25.2
scipy==1.15.3
sympy==1.14.0
tifffile==2025.5.10
timm==1.0.22
tokenizers==0.22.1
torch==2.9.1
torchaudio==2.9.1
torchvision==0.24.1
tqdm==4.67.1
transformers==4.57.3
triton==3.5.1
typing_extensions==4.15.0
urllib3==2.5.0
wcwidth==0.2.14
